refactor: Router
Created two classes Route and Render. The general idea is Route is
responsible for resolving, redirecting, checking 404, etc.
It wlll also call get_context to generate the context for that route.
A Route object will contain all the meta-data that is required to render
that route.
Render takes a Route object and renders it. It will also cache it if
the route is cacheable.